@@ -1,92 +1,44 @@
-buildscript {
-    repositories {
-        maven { url 'https://maven.fabric.io/public' }
-    }
-    dependencies {
-        classpath 'io.fabric.tools:gradle:1.+'
-    }
-}
 apply plugin: 'com.android.application'
-apply plugin: 'io.fabric'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
     defaultConfig {
-        applicationId "com.eduvanzapplication"
+        applicationId "com.smartbizz"
         minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 134
-        versionName "5.3.4"
+        targetSdkVersion 29
+        versionCode 2
+        versionName "1.0.1"
         vectorDrawables.useSupportLibrary = true
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
-
-//        Friendly-Score
-        manifestPlaceholders = [
-                'appAuthRedirectScheme': 'com.eduvanzapplication'
-        ]
-
-//        ndk{
-//            abiFilters "arm64-v8a", "armeabi-v7a", "x86", "x86_64"
-//        }
-
-        renderscriptTargetApi 21
-        renderscriptSupportModeEnabled true
+//Samarth
     }
 
-    packagingOptions {
-        pickFirst 'lib/armeabi-v7a/libRSSupport.so'
-        pickFirst 'lib/arm64-v8a/librsjni.so'
-
-        pickFirst 'lib/x86_64/libopencv_java3.so'
-        pickFirst 'lib/x86/libopencv_java3.so'
-        pickFirst 'lib/armeabi-v7a/libopencv_java3.so'
-        pickFirst 'lib/arm64-v8a/libopencv_java3.so'
-    }
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    lintOptions {
-        disable 'MissingTranslation'
-        checkReleaseBuilds false
-        abortOnError false
-    }
-    configurations {
-        all*.exclude group: 'com.android.volley'
-    }
+
+//    lintOptions {
+//        disable 'MissingTranslation'
+//        checkReleaseBuilds false
+//        abortOnError false
+//    }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
 
-    splits {
-// Configures multiple APKs based on ABI.
-        abi {
-// Enables building multiple APKs per ABI.
-            enable true
-
-// By default all ABIs are included
-// so use reset()
-            reset()
-
-// Specifies a list of ABIs that Gradle should create APKs for.
-            include "arm64-v8a", "armeabi-v7a", "x86", "x86_64"
-
-// Specifies that we want to also generate a universal APK that includes all ABIs.
-            universalApk true
-        }
-    }
 }
 
 repositories {
     mavenCentral()
     maven { url 'https://s3.amazonaws.com/repo.commonsware.com' }
     maven { url 'https://jitpack.io' }
-    maven { url 'https://maven.fabric.io/public' }
 }
 
 dependencies {
