@@ -1,7 +1,7 @@
 package com.yuyakaido.android.cardstackview.internal;
 
-import android.support.annotation.Nullable;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.yuyakaido.android.cardstackview.CardStackLayoutManager;
 
