@@ -1,6 +1,6 @@
 package com.yuyakaido.android.cardstackview.internal;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 import com.yuyakaido.android.cardstackview.CardStackLayoutManager;
