@@ -1,9 +1,9 @@
 package com.yuyakaido.android.cardstackview.internal;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SnapHelper;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.SnapHelper;
 import android.view.View;
 
 import com.yuyakaido.android.cardstackview.CardStackLayoutManager;
